StreamlitUI/requirements.txt
PyPDF2
python-docx
streamlit
requests